Feature: Allow running a command after successful cronjob.
Using the `cron_success_command` parameter for `certonly`, you can
run commands when the renewal command was successful, e.g.
reload the webserver to use the renewed certs.

Not sure about letsencrypt return values, so the command might be
executed even if the cert is unchanged, doesn't matter for this
usecase though.